Moved TimeScale to separate class
