+++ src/main/java/org/radarcns/webapp/resource/AggregatedDataPointsEndPoint.java
@@ -32,7 +32,7 @@ import org.radarcns.webapp.filter.Authenticated;
 import org.radarcns.webapp.param.DataAggregateParam;
 import org.radarcns.webapp.param.InstantParam;
 import org.radarcns.webapp.param.TimeScaleParser;
-import org.radarcns.webapp.param.TimeScaleParser.TimeScale;
+import org.radarcns.webapp.param.TimeScale;
 import org.radarcns.webapp.validation.Alphanumeric;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
